+++ refgenieserver/helpers.py
@@ -60,20 +60,38 @@ def build_parser():
         type=int,
         help="The port the webserver should be run on.", default=DEFAULT_PORT)
     sps["archive"].add_argument(
+        "--genomes_desc",
+        dest="genomes_desc",
+        type=str,
+        default=None,
+        help="Path to a CSV file with genomes descriptions. Format: genome_name, genome description")
+    sps["archive"].add_argument(
         "-f", "--force",
         action="store_true",
         dest="force",
         help="whether the server file tree should be rebuilt even if exists")
     sps["archive"].add_argument(
-        "-g", "--genome",
-        dest="genome",
-        nargs=1,
-        type=str,
-        help="request a specific genome build")
+        "-r", "--remove",
+        action="store_true",
+        dest="remove",
+        help="Remove selected genome, genome/asset or genome/asset:tag")
     sps["archive"].add_argument(
-        "-a", "--asset",
-        type=str,
-        nargs="*",
-        dest="asset",
-        help="request a specific asset build")
+        "asset_registry_paths", metavar="asset-registry-paths", type=str, nargs='*',
+        help="One or more registry path strings that identify assets, e.g. hg38/fasta:tag")
     return parser
+
+
+def preprocess_attrs(attrs):
+    """
+    Based on the CHANGED_KEYS mapping (new_key:old_key), rename the keys in the provided one
+
+    :param yacman.yacman.YacAttMap attrs: mapping to process
+    :return yacman.yacman.YacAttMap: mapping with renamed key names
+    """
+    from copy import deepcopy
+    attrs_cpy = deepcopy(attrs)
+    for new_key in CHANGED_KEYS.keys():
+        if new_key in attrs_cpy:
+            attrs_cpy[CHANGED_KEYS[new_key]] = attrs_cpy[new_key]
+            del attrs_cpy[new_key]
+    return attrs_cpy